Extract legacy text helpers
